fix: consolidate Python API tests by removing redundant test file
Removed basic test_simple_api.py and renamed comprehensive test_simple_api_real.py
to test_simple_api.py. The "real" tests actually validate against Fortran behavior
with proper constant checking, parameter validation, and robustness testing.
